+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.0.4] 
 - fix gelf handler race condition issue
+- add standard field type
 
 ## [1.0.3] 2020-03-13
 - use slice fields to improve performance